+++ src/app/alpaca/page.tsx
@@ -1,6 +1,6 @@
 import { useTranslations } from 'next-intl';
 import AlpacaAssets from "@/components/alpaca/AlpacaAssets";
-import AlpacaHistoryChart from '@/components/alpaca/AlpacaHistoryChart';
+import AlpacaHistoryChart from '@/components/alpaca/history-charts/AlpacaHistoryChart';
 import * as React from 'react';
 import "../../app/globals.css";
 
@@ -11,10 +11,10 @@ export default function AlpacaDashboard() {
   return (
 
     <div className="flex h-full items-stretch">
-      <div className='overflow-hidden w-1/3 p-2 pr-1'>
+      <div className='overflow-hidden w-1/4 p-2 pr-1'>
         <AlpacaAssets />
       </div>
-      <div className='overflow-hidden w-2/3 p-2 pl-1'>
+      <div className='overflow-hidden w-3/4 p-2 pl-1'>
         <AlpacaHistoryChart />
       </div>
     </div>