+++ src/components/navigation.tsx
@@ -6,7 +6,6 @@ import Link from 'next/link';
 import HomeIcon from '@mui/icons-material/Home';
 import DashboardIcon from '@mui/icons-material/Dashboard';
 import BarChartIcon from '@mui/icons-material/BarChart';
-import DoubleArrowIcon from '@mui/icons-material/DoubleArrow';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 
@@ -24,21 +23,31 @@ const Navigation = () => {
       <ul className="space-y-4  py-4  border-t border-b border-slate-400">
         <li>
           <Link href="/" className="grid grid-cols-[20px_auto] items-center gap-2" title="Home">
-              <HomeIcon className="h-6 w-6 text-blue-500" />
-              {expanded && <span className='text-base'>{t('home')}</span>}
+            <HomeIcon className="h-6 w-6 text-blue-500" />
+            {expanded && <span className='text-base'>{t('home')}</span>}
           </Link>
         </li>
         <li>
           <Link href="/dashboard" className="grid grid-cols-[20px_auto] items-center gap-2" title="Dashboard">
-              <DashboardIcon className="h-6 w-6 text-blue-500" />
-              {expanded && <span className='text-base'>{t('dashboard')}</span>}
+            <DashboardIcon className="h-6 w-6 text-blue-500" />
+            {expanded && <span className='text-base'>{t('dashboard')}</span>}
           </Link>
         </li>
         <li>
           <Link href="/alpaca" className="grid grid-cols-[20px_auto] items-center gap-2" title="Alpaca">
-              <BarChartIcon className="h-6 w-6 text-blue-500" />
-              {expanded && <span className='text-base'>{t('alpaca')}</span>}
+            <BarChartIcon className="h-6 w-6 text-blue-500" />
+            {expanded && <span className='text-base'>{t('alpaca')}</span>}
           </Link>
+          <div className="pl-5 border border-lime-600">
+            <ul>
+              <li className="text-white">
+              <Link href="/alpaca" className="" title="Alpaca">
+
+                {t('alpaca')}
+              </Link>
+            </li></ul>
+          </div>
+
         </li>
       </ul>
       <button onClick={toggleView} className="flex justify-end items-end my-4">