+++ Release-Notes.md
@@ -1,5 +1,15 @@
 # BN TRADING APP
 
+## Version 2024.8.1
+
+Loader integration
+Input form for chart parameters
+Menu rework
+rework useSWR reload and api implementation via route.ts
+
 ## Version 2024.7.1
 
-Layout
+Layout tailwind based (flex)
+API server/client
+Highcharts integration
+Asset list